requirements.txt
streamlit
google-genai
pypdf
python-docx
python-dotenv
